editable activities (db)
